% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HAPT.R
\docType{data}
\name{HAPT}
\alias{HAPT}
\title{Human Activity Recognition Dataset}
\format{\describe{
\item{\samp{1}{X0.043579674}}{a numeric vector}
\item{\samp{2}{X.0.005970221}}{a numeric vector}
\item{\samp{3}{X.0.035054344}}{a numeric vector}
\item{\samp{4}{X.0.99538116}}{a numeric vector}
\item{\samp{5}{X.0.988365863}}{a numeric vector}
\item{\samp{6}{X.0.937382005}}{a numeric vector}
\item{\samp{7}{X.0.995007045}}{a numeric vector}
\item{\samp{8}{X.0.988815577}}{a numeric vector}
\item{\samp{9}{X.0.953325201}}{a numeric vector}
\item{\samp{10}{X.0.794796369}}{a numeric vector}
\item{\samp{11}{X.0.744892817}}{a numeric vector}
\item{\samp{12}{X.0.648447245}}{a numeric vector}
\item{\samp{13}{X0.841795574}}{a numeric vector}
\item{\samp{14}{X0.708440184}}{a numeric vector}
\item{\samp{15}{X0.651716486}}{a numeric vector}
\item{\samp{16}{X.0.975752033}}{a numeric vector}
\item{\samp{17}{X.0.999950218}}{a numeric vector}
\item{\samp{18}{X.0.999888368}}{a numeric vector}
\item{\samp{19}{X.0.998013666}}{a numeric vector}
\item{\samp{20}{X.0.993999294}}{a numeric vector}
\item{\samp{21}{X.0.991979622}}{a numeric vector}
\item{\samp{22}{X.0.970970054}}{a numeric vector}
\item{\samp{23}{X.0.547095381}}{a numeric vector}
\item{\samp{24}{X.0.70097364}}{a numeric vector}
\item{\samp{25}{X.0.622697109}}{a numeric vector}
\item{\samp{26}{X0.921884153}}{a numeric vector}
\item{\samp{27}{X.0.719482881}}{a numeric vector}
\item{\samp{28}{X0.342168007}}{a numeric vector}
\item{\samp{29}{X.0.161317646}}{a numeric vector}
\item{\samp{30}{X0.266048969}}{a numeric vector}
\item{\samp{31}{X.0.274351454}}{a numeric vector}
\item{\samp{32}{X0.26720509}}{a numeric vector}
\item{\samp{33}{X.0.020958427}}{a numeric vector}
\item{\samp{34}{X0.382610239}}{a numeric vector}
\item{\samp{35}{X.0.501748058}}{a numeric vector}
\item{\samp{36}{X0.51246321}}{a numeric vector}
\item{\samp{37}{X.0.206336868}}{a numeric vector}
\item{\samp{38}{X0.376778004}}{a numeric vector}
\item{\samp{39}{X0.435172317}}{a numeric vector}
\item{\samp{40}{X0.66019853}}{a numeric vector}
\item{\samp{41}{X0.960050684}}{a numeric vector}
\item{\samp{42}{X.0.135938708}}{a numeric vector}
\item{\samp{43}{X0.11555554}}{a numeric vector}
\item{\samp{44}{X.0.988134472}}{a numeric vector}
\item{\samp{45}{X.0.982692844}}{a numeric vector}
\item{\samp{46}{X.0.919722675}}{a numeric vector}
\item{\samp{47}{X.0.988361912}}{a numeric vector}
\item{\samp{48}{X.0.985522958}}{a numeric vector}
\item{\samp{49}{X.0.931833779}}{a numeric vector}
\item{\samp{50}{X0.892054508}}{a numeric vector}
\item{\samp{51}{X.0.161350449}}{a numeric vector}
\item{\samp{52}{X0.124659771}}{a numeric vector}
\item{\samp{53}{X0.977436313}}{a numeric vector}
\item{\samp{54}{X.0.121512811}}{a numeric vector}
\item{\samp{55}{X0.056114283}}{a numeric vector}
\item{\samp{56}{X.0.374464221}}{a numeric vector}
\item{\samp{57}{X0.891821485}}{a numeric vector}
\item{\samp{58}{X.0.970835551}}{a numeric vector}
\item{\samp{59}{X.0.975510738}}{a numeric vector}
\item{\samp{60}{X.0.989140808}}{a numeric vector}
\item{\samp{61}{X.0.990444311}}{a numeric vector}
\item{\samp{62}{X.0.948886957}}{a numeric vector}
\item{\samp{63}{X.1}}{a numeric vector}
\item{\samp{64}{X.1.1}}{a numeric vector}
\item{\samp{65}{X0.034248559}}{a numeric vector}
\item{\samp{66}{X.0.593146212}}{a numeric vector}
\item{\samp{67}{X0.59463587}}{a numeric vector}
\item{\samp{68}{X.0.596439899}}{a numeric vector}
\item{\samp{69}{X0.598560307}}{a numeric vector}
\item{\samp{70}{X.0.712209102}}{a numeric vector}
\item{\samp{71}{X0.709630397}}{a numeric vector}
\item{\samp{72}{X.0.707484484}}{a numeric vector}
\item{\samp{73}{X0.705000498}}{a numeric vector}
\item{\samp{74}{X.0.992117053}}{a numeric vector}
\item{\samp{75}{X0.992646866}}{a numeric vector}
\item{\samp{76}{X.0.992885426}}{a numeric vector}
\item{\samp{77}{X0.991747853}}{a numeric vector}
\item{\samp{78}{X0.570268756}}{a numeric vector}
\item{\samp{79}{X0.439065091}}{a numeric vector}
\item{\samp{80}{X0.986912961}}{a numeric vector}
\item{\samp{81}{X0.077996345}}{a numeric vector}
\item{\samp{82}{X0.005000803}}{a numeric vector}
\item{\samp{83}{X.0.097789591}}{a numeric vector}
\item{\samp{84}{X.0.99351906}}{a numeric vector}
\item{\samp{85}{X.0.988359988}}{a numeric vector}
\item{\samp{86}{X.0.993574969}}{a numeric vector}
\item{\samp{87}{X.0.994437381}}{a numeric vector}
\item{\samp{88}{X.0.986206639}}{a numeric vector}
\item{\samp{89}{X.0.992818354}}{a numeric vector}
\item{\samp{90}{X.0.985180095}}{a numeric vector}
\item{\samp{91}{X.0.991994232}}{a numeric vector}
\item{\samp{92}{X.0.993118865}}{a numeric vector}
\item{\samp{93}{X0.98983471}}{a numeric vector}
\item{\samp{94}{X0.991956863}}{a numeric vector}
\item{\samp{95}{X0.990519204}}{a numeric vector}
\item{\samp{96}{X.0.993658164}}{a numeric vector}
\item{\samp{97}{X.0.999934868}}{a numeric vector}
\item{\samp{98}{X.0.999820446}}{a numeric vector}
\item{\samp{99}{X.0.999878465}}{a numeric vector}
\item{\samp{100}{X.0.994406779}}{a numeric vector}
\item{\samp{101}{X.0.986024868}}{a numeric vector}
\item{\samp{102}{X.0.988697635}}{a numeric vector}
\item{\samp{103}{X.0.819949251}}{a numeric vector}
\item{\samp{104}{X.0.793046449}}{a numeric vector}
\item{\samp{105}{X.0.863145219}}{a numeric vector}
\item{\samp{106}{X0.948026335}}{a numeric vector}
\item{\samp{107}{X.0.334807052}}{a numeric vector}
\item{\samp{108}{X0.714201993}}{a numeric vector}
\item{\samp{109}{X0.49868609}}{a numeric vector}
\item{\samp{110}{X0.296680795}}{a numeric vector}
\item{\samp{111}{X.0.18951272}}{a numeric vector}
\item{\samp{112}{X0.310911634}}{a numeric vector}
\item{\samp{113}{X0.237110093}}{a numeric vector}
\item{\samp{114}{X0.410936621}}{a numeric vector}
\item{\samp{115}{X.0.509475111}}{a numeric vector}
\item{\samp{116}{X0.475573642}}{a numeric vector}
\item{\samp{117}{X.0.352207366}}{a numeric vector}
\item{\samp{118}{X.0.394630885}}{a numeric vector}
\item{\samp{119}{X.0.01679047}}{a numeric vector}
\item{\samp{120}{X.0.15823668}}{a numeric vector}
\item{\samp{121}{X.0.006100849}}{a numeric vector}
\item{\samp{122}{X0.007514728}}{a numeric vector}
\item{\samp{123}{X.0.063861644}}{a numeric vector}
\item{\samp{124}{X.0.985312105}}{a numeric vector}
\item{\samp{125}{X.0.976607569}}{a numeric vector}
\item{\samp{126}{X.0.99331883}}{a numeric vector}
\item{\samp{127}{X.0.984589148}}{a numeric vector}
\item{\samp{128}{X.0.976326479}}{a numeric vector}
\item{\samp{129}{X.0.993341427}}{a numeric vector}
\item{\samp{130}{X.0.89756307}}{a numeric vector}
\item{\samp{131}{X.0.936675984}}{a numeric vector}
\item{\samp{132}{X.0.762644342}}{a numeric vector}
\item{\samp{133}{X0.854297167}}{a numeric vector}
\item{\samp{134}{X0.895031593}}{a numeric vector}
\item{\samp{135}{X0.83084054}}{a numeric vector}
\item{\samp{136}{X.0.967121622}}{a numeric vector}
\item{\samp{137}{X.0.999578211}}{a numeric vector}
\item{\samp{138}{X.0.999354327}}{a numeric vector}
\item{\samp{139}{X.0.999818097}}{a numeric vector}
\item{\samp{140}{X.0.984399532}}{a numeric vector}
\item{\samp{141}{X.0.978576803}}{a numeric vector}
\item{\samp{142}{X.0.993052695}}{a numeric vector}
\item{\samp{143}{X0.082631682}}{a numeric vector}
\item{\samp{144}{X0.174515435}}{a numeric vector}
\item{\samp{145}{X.0.218436995}}{a numeric vector}
\item{\samp{146}{X0.320926125}}{a numeric vector}
\item{\samp{147}{X.0.452603567}}{a numeric vector}
\item{\samp{148}{X0.466132542}}{a numeric vector}
\item{\samp{149}{X.0.20363231}}{a numeric vector}
\item{\samp{150}{X1}}{a numeric vector}
\item{\samp{151}{X.1.2}}{a numeric vector}
\item{\samp{152}{X0.352275964}}{a numeric vector}
\item{\samp{153}{X0.448398409}}{a numeric vector}
\item{\samp{154}{X0.757793127}}{a numeric vector}
\item{\samp{155}{X.0.782464781}}{a numeric vector}
\item{\samp{156}{X0.70372469}}{a numeric vector}
\item{\samp{157}{X.0.399106532}}{a numeric vector}
\item{\samp{158}{X0.71227414}}{a numeric vector}
\item{\samp{159}{X0.135603878}}{a numeric vector}
\item{\samp{160}{X0.302777347}}{a numeric vector}
\item{\samp{161}{X.0.0991674}}{a numeric vector}
\item{\samp{162}{X.0.055517369}}{a numeric vector}
\item{\samp{163}{X.0.061985797}}{a numeric vector}
\item{\samp{164}{X.0.991918262}}{a numeric vector}
\item{\samp{165}{X.0.992519273}}{a numeric vector}
\item{\samp{166}{X.0.99235236}}{a numeric vector}
\item{\samp{167}{X.0.991961469}}{a numeric vector}
\item{\samp{168}{X.0.994949185}}{a numeric vector}
\item{\samp{169}{X.0.992834951}}{a numeric vector}
\item{\samp{170}{X.0.991765744}}{a numeric vector}
\item{\samp{171}{X.0.986564774}}{a numeric vector}
\item{\samp{172}{X.0.991893633}}{a numeric vector}
\item{\samp{173}{X0.994225213}}{a numeric vector}
\item{\samp{174}{X0.99176104}}{a numeric vector}
\item{\samp{175}{X0.989351947}}{a numeric vector}
\item{\samp{176}{X.0.994522505}}{a numeric vector}
\item{\samp{177}{X.0.999936809}}{a numeric vector}
\item{\samp{178}{X.0.999953503}}{a numeric vector}
\item{\samp{179}{X.0.999924694}}{a numeric vector}
\item{\samp{180}{X.0.992214405}}{a numeric vector}
\item{\samp{181}{X.0.996938049}}{a numeric vector}
\item{\samp{182}{X.0.992583909}}{a numeric vector}
\item{\samp{183}{X.0.589850963}}{a numeric vector}
\item{\samp{184}{X.0.709836492}}{a numeric vector}
\item{\samp{185}{X.0.572106861}}{a numeric vector}
\item{\samp{186}{X0.374661331}}{a numeric vector}
\item{\samp{187}{X.0.306094434}}{a numeric vector}
\item{\samp{188}{X0.335505508}}{a numeric vector}
\item{\samp{189}{X.0.079307172}}{a numeric vector}
\item{\samp{190}{X0.989559599}}{a numeric vector}
\item{\samp{191}{X.0.454970093}}{a numeric vector}
\item{\samp{192}{X0.384435155}}{a numeric vector}
\item{\samp{193}{X0.307638223}}{a numeric vector}
\item{\samp{194}{X0.936231821}}{a numeric vector}
\item{\samp{195}{X.0.666281656}}{a numeric vector}
\item{\samp{196}{X0.772201678}}{a numeric vector}
\item{\samp{197}{X0.529584895}}{a numeric vector}
\item{\samp{198}{X.0.147221207}}{a numeric vector}
\item{\samp{199}{X.0.094462012}}{a numeric vector}
\item{\samp{200}{X0.351206262}}{a numeric vector}
\item{\samp{201}{X.0.969546337}}{a numeric vector}
\item{\samp{202}{X.0.961539739}}{a numeric vector}
\item{\samp{203}{X.0.967645839}}{a numeric vector}
\item{\samp{204}{X.0.9472362}}{a numeric vector}
\item{\samp{205}{X.0.994174403}}{a numeric vector}
\item{\samp{206}{X.0.969546337.1}}{a numeric vector}
\item{\samp{207}{X.0.999150002}}{a numeric vector}
\item{\samp{208}{X.0.970525396}}{a numeric vector}
\item{\samp{209}{X.0.21961956}}{a numeric vector}
\item{\samp{210}{X.0.124072048}}{a numeric vector}
\item{\samp{211}{X.0.058240845}}{a numeric vector}
\item{\samp{212}{X0.174306353}}{a numeric vector}
\item{\samp{213}{X0.201620464}}{a numeric vector}
\item{\samp{214}{X.0.969546337.2}}{a numeric vector}
\item{\samp{215}{X.0.961539739.1}}{a numeric vector}
\item{\samp{216}{X.0.967645839.1}}{a numeric vector}
\item{\samp{217}{X.0.9472362.1}}{a numeric vector}
\item{\samp{218}{X.0.994174403.1}}{a numeric vector}
\item{\samp{219}{X.0.969546337.3}}{a numeric vector}
\item{\samp{220}{X.0.999150002.1}}{a numeric vector}
\item{\samp{221}{X.0.970525396.1}}{a numeric vector}
\item{\samp{222}{X.0.21961956.1}}{a numeric vector}
\item{\samp{223}{X.0.124072048.1}}{a numeric vector}
\item{\samp{224}{X.0.058240845.1}}{a numeric vector}
\item{\samp{225}{X0.174306353.1}}{a numeric vector}
\item{\samp{226}{X0.201620464.1}}{a numeric vector}
\item{\samp{227}{X.0.994088281}}{a numeric vector}
\item{\samp{228}{X.0.994309239}}{a numeric vector}
\item{\samp{229}{X.0.994421392}}{a numeric vector}
\item{\samp{230}{X.0.992435558}}{a numeric vector}
\item{\samp{231}{X.0.991208465}}{a numeric vector}
\item{\samp{232}{X.0.994088281.1}}{a numeric vector}
\item{\samp{233}{X.0.999901007}}{a numeric vector}
\item{\samp{234}{X.0.992971863}}{a numeric vector}
\item{\samp{235}{X.0.863283735}}{a numeric vector}
\item{\samp{236}{X0.30327694}}{a numeric vector}
\item{\samp{237}{X.0.26616858}}{a numeric vector}
\item{\samp{238}{X.0.199724191}}{a numeric vector}
\item{\samp{239}{X0.067559792}}{a numeric vector}
\item{\samp{240}{X.0.968904091}}{a numeric vector}
\item{\samp{241}{X.0.964335179}}{a numeric vector}
\item{\samp{242}{X.0.959637641}}{a numeric vector}
\item{\samp{243}{X.0.975108238}}{a numeric vector}
\item{\samp{244}{X.0.99500898}}{a numeric vector}
\item{\samp{245}{X.0.968904091.1}}{a numeric vector}
\item{\samp{246}{X.0.99928627}}{a numeric vector}
\item{\samp{247}{X.0.954863924}}{a numeric vector}
\item{\samp{248}{X0.189420069}}{a numeric vector}
\item{\samp{249}{X0.556826973}}{a numeric vector}
\item{\samp{250}{X.0.830368613}}{a numeric vector}
\item{\samp{251}{X0.302198244}}{a numeric vector}
\item{\samp{252}{X0.435743292}}{a numeric vector}
\item{\samp{253}{X.0.994299894}}{a numeric vector}
\item{\samp{254}{X.0.991367611}}{a numeric vector}
\item{\samp{255}{X.0.993142976}}{a numeric vector}
\item{\samp{256}{X.0.988935628}}{a numeric vector}
\item{\samp{257}{X.0.993486031}}{a numeric vector}
\item{\samp{258}{X.0.994299894.1}}{a numeric vector}
\item{\samp{259}{X.0.999949174}}{a numeric vector}
\item{\samp{260}{X.0.994547178}}{a numeric vector}
\item{\samp{261}{X.0.619767633}}{a numeric vector}
\item{\samp{262}{X0.329123565}}{a numeric vector}
\item{\samp{263}{X.0.149681742}}{a numeric vector}
\item{\samp{264}{X.0.218394068}}{a numeric vector}
\item{\samp{265}{X.0.037588387}}{a numeric vector}
\item{\samp{266}{X.0.994809834}}{a numeric vector}
\item{\samp{267}{X.0.982984099}}{a numeric vector}
\item{\samp{268}{X.0.939455343}}{a numeric vector}
\item{\samp{269}{X.0.995272004}}{a numeric vector}
\item{\samp{270}{X.0.989848258}}{a numeric vector}
\item{\samp{271}{X.0.940334491}}{a numeric vector}
\item{\samp{272}{X.0.996824354}}{a numeric vector}
\item{\samp{273}{X.0.984703069}}{a numeric vector}
\item{\samp{274}{X.0.933408785}}{a numeric vector}
\item{\samp{275}{X.0.993663737}}{a numeric vector}
\item{\samp{276}{X.0.990217401}}{a numeric vector}
\item{\samp{277}{X.0.942331996}}{a numeric vector}
\item{\samp{278}{X.0.993952664}}{a numeric vector}
\item{\samp{279}{X.0.993447891}}{a numeric vector}
\item{\samp{280}{X.0.923441387}}{a numeric vector}
\item{\samp{281}{X.0.974732706}}{a numeric vector}
\item{\samp{282}{X.0.999968834}}{a numeric vector}
\item{\samp{283}{X.0.999850408}}{a numeric vector}
\item{\samp{284}{X.0.997308392}}{a numeric vector}
\item{\samp{285}{X.0.996767586}}{a numeric vector}
\item{\samp{286}{X.0.989708889}}{a numeric vector}
\item{\samp{287}{X.0.990807294}}{a numeric vector}
\item{\samp{288}{X.0.94635692}}{a numeric vector}
\item{\samp{289}{X.0.904747763}}{a numeric vector}
\item{\samp{290}{X.0.591302475}}{a numeric vector}
\item{\samp{291}{X.1.3}}{a numeric vector}
\item{\samp{292}{X.1.4}}{a numeric vector}
\item{\samp{293}{X.1.5}}{a numeric vector}
\item{\samp{294}{X0.362081644}}{a numeric vector}
\item{\samp{295}{X0.159191141}}{a numeric vector}
\item{\samp{296}{X0.04834979}}{a numeric vector}
\item{\samp{297}{X0.141770079}}{a numeric vector}
\item{\samp{298}{X.0.151402429}}{a numeric vector}
\item{\samp{299}{X.0.260648523}}{a numeric vector}
\item{\samp{300}{X.0.551541897}}{a numeric vector}
\item{\samp{301}{X0.253213642}}{a numeric vector}
\item{\samp{302}{X.0.019604287}}{a numeric vector}
\item{\samp{303}{X.0.999962575}}{a numeric vector}
\item{\samp{304}{X.0.999987355}}{a numeric vector}
\item{\samp{305}{X.0.999979068}}{a numeric vector}
\item{\samp{306}{X.0.999962451}}{a numeric vector}
\item{\samp{307}{X.0.999932224}}{a numeric vector}
\item{\samp{308}{X.0.999726224}}{a numeric vector}
\item{\samp{309}{X.0.999670333}}{a numeric vector}
\item{\samp{310}{X.0.999985825}}{a numeric vector}
\item{\samp{311}{X.0.999968239}}{a numeric vector}
\item{\samp{312}{X.0.99997686}}{a numeric vector}
\item{\samp{313}{X.0.999868536}}{a numeric vector}
\item{\samp{314}{X.0.999776093}}{a numeric vector}
\item{\samp{315}{X.0.999971375}}{a numeric vector}
\item{\samp{316}{X.0.999917656}}{a numeric vector}
\item{\samp{317}{X.0.999885011}}{a numeric vector}
\item{\samp{318}{X.0.999860464}}{a numeric vector}
\item{\samp{319}{X.0.999868935}}{a numeric vector}
\item{\samp{320}{X.0.999862134}}{a numeric vector}
\item{\samp{321}{X.0.999745894}}{a numeric vector}
\item{\samp{322}{X.0.999721346}}{a numeric vector}
\item{\samp{323}{X.0.9994931}}{a numeric vector}
\item{\samp{324}{X.0.999813634}}{a numeric vector}
\item{\samp{325}{X.0.999860836}}{a numeric vector}
\item{\samp{326}{X.0.999838264}}{a numeric vector}
\item{\samp{327}{X.0.999741864}}{a numeric vector}
\item{\samp{328}{X.0.999615143}}{a numeric vector}
\item{\samp{329}{X.0.999855752}}{a numeric vector}
\item{\samp{330}{X.0.999838629}}{a numeric vector}
\item{\samp{331}{X.0.997422354}}{a numeric vector}
\item{\samp{332}{X.0.999470578}}{a numeric vector}
\item{\samp{333}{X.0.999662043}}{a numeric vector}
\item{\samp{334}{X.0.999642298}}{a numeric vector}
\item{\samp{335}{X.0.999296892}}{a numeric vector}
\item{\samp{336}{X.0.997892219}}{a numeric vector}
\item{\samp{337}{X.0.995932257}}{a numeric vector}
\item{\samp{338}{X.0.995513968}}{a numeric vector}
\item{\samp{339}{X.0.99736675}}{a numeric vector}
\item{\samp{340}{X.0.999685082}}{a numeric vector}
\item{\samp{341}{X.0.998917794}}{a numeric vector}
\item{\samp{342}{X.0.995667147}}{a numeric vector}
\item{\samp{343}{X.0.997347905}}{a numeric vector}
\item{\samp{344}{X.0.999453818}}{a numeric vector}
\item{\samp{345}{X.0.992332448}}{a numeric vector}
\item{\samp{346}{X.0.987169905}}{a numeric vector}
\item{\samp{347}{X.0.989643691}}{a numeric vector}
\item{\samp{348}{X.0.995820679}}{a numeric vector}
\item{\samp{349}{X.0.990936306}}{a numeric vector}
\item{\samp{350}{X.0.996462312}}{a numeric vector}
\item{\samp{351}{X.0.993805471}}{a numeric vector}
\item{\samp{352}{X.0.990518692}}{a numeric vector}
\item{\samp{353}{X.0.997124222}}{a numeric vector}
\item{\samp{354}{X.0.996740158}}{a numeric vector}
\item{\samp{355}{X.0.991975155}}{a numeric vector}
\item{\samp{356}{X.0.993156727}}{a numeric vector}
\item{\samp{357}{X.0.998353298}}{a numeric vector}
\item{\samp{358}{X.0.991107714}}{a numeric vector}
\item{\samp{359}{X.0.959896403}}{a numeric vector}
\item{\samp{360}{X.0.990984417}}{a numeric vector}
\item{\samp{361}{X.0.999934753}}{a numeric vector}
\item{\samp{362}{X.0.99982048}}{a numeric vector}
\item{\samp{363}{X.0.999884492}}{a numeric vector}
\item{\samp{364}{X.0.992164178}}{a numeric vector}
\item{\samp{365}{X.0.991419598}}{a numeric vector}
\item{\samp{366}{X.0.995502344}}{a numeric vector}
\item{\samp{367}{X.1.6}}{a numeric vector}
\item{\samp{368}{X.1.7}}{a numeric vector}
\item{\samp{369}{X.1.8}}{a numeric vector}
\item{\samp{370}{X1.1}}{a numeric vector}
\item{\samp{371}{X.0.24}}{a numeric vector}
\item{\samp{372}{X.1.9}}{a numeric vector}
\item{\samp{373}{X0.700792717}}{a numeric vector}
\item{\samp{374}{X0.212389878}}{a numeric vector}
\item{\samp{375}{X0.36443644}}{a numeric vector}
\item{\samp{376}{X.0.717928666}}{a numeric vector}
\item{\samp{377}{X.0.920370675}}{a numeric vector}
\item{\samp{378}{X.0.643843696}}{a numeric vector}
\item{\samp{379}{X.0.936310054}}{a numeric vector}
\item{\samp{380}{X.0.483802686}}{a numeric vector}
\item{\samp{381}{X.0.818679801}}{a numeric vector}
\item{\samp{382}{X.0.999986577}}{a numeric vector}
\item{\samp{383}{X.0.999980394}}{a numeric vector}
\item{\samp{384}{X.0.999974784}}{a numeric vector}
\item{\samp{385}{X.0.999954986}}{a numeric vector}
\item{\samp{386}{X.0.999918605}}{a numeric vector}
\item{\samp{387}{X.0.999640039}}{a numeric vector}
\item{\samp{388}{X.0.999484091}}{a numeric vector}
\item{\samp{389}{X.0.999960866}}{a numeric vector}
\item{\samp{390}{X.0.999982549}}{a numeric vector}
\item{\samp{391}{X.0.999971294}}{a numeric vector}
\item{\samp{392}{X.0.999816816}}{a numeric vector}
\item{\samp{393}{X.0.999487249}}{a numeric vector}
\item{\samp{394}{X.0.999979215}}{a numeric vector}
\item{\samp{395}{X.0.999853126}}{a numeric vector}
\item{\samp{396}{X.0.99993261}}{a numeric vector}
\item{\samp{397}{X.0.999899927}}{a numeric vector}
\item{\samp{398}{X.0.999824442}}{a numeric vector}
\item{\samp{399}{X.0.999857117}}{a numeric vector}
\item{\samp{400}{X.0.999732888}}{a numeric vector}
\item{\samp{401}{X.0.999727183}}{a numeric vector}
\item{\samp{402}{X.0.999567073}}{a numeric vector}
\item{\samp{403}{X.0.999765245}}{a numeric vector}
\item{\samp{404}{X.0.999900254}}{a numeric vector}
\item{\samp{405}{X.0.999812827}}{a numeric vector}
\item{\samp{406}{X.0.999705237}}{a numeric vector}
\item{\samp{407}{X.0.999596081}}{a numeric vector}
\item{\samp{408}{X.0.999852161}}{a numeric vector}
\item{\samp{409}{X.0.999822101}}{a numeric vector}
\item{\samp{410}{X.0.999398885}}{a numeric vector}
\item{\samp{411}{X.0.999763402}}{a numeric vector}
\item{\samp{412}{X.0.999958458}}{a numeric vector}
\item{\samp{413}{X.0.999943843}}{a numeric vector}
\item{\samp{414}{X.0.99982466}}{a numeric vector}
\item{\samp{415}{X.0.999813509}}{a numeric vector}
\item{\samp{416}{X.0.998775859}}{a numeric vector}
\item{\samp{417}{X.0.998577775}}{a numeric vector}
\item{\samp{418}{X.0.999617411}}{a numeric vector}
\item{\samp{419}{X.0.999983904}}{a numeric vector}
\item{\samp{420}{X.0.999832422}}{a numeric vector}
\item{\samp{421}{X.0.99867621}}{a numeric vector}
\item{\samp{422}{X.0.999844338}}{a numeric vector}
\item{\samp{423}{X.0.9999223}}{a numeric vector}
\item{\samp{424}{X.0.986589704}}{a numeric vector}
\item{\samp{425}{X.0.981771013}}{a numeric vector}
\item{\samp{426}{X.0.989355857}}{a numeric vector}
\item{\samp{427}{X.0.985027674}}{a numeric vector}
\item{\samp{428}{X.0.97390699}}{a numeric vector}
\item{\samp{429}{X.0.994979154}}{a numeric vector}
\item{\samp{430}{X.0.987339664}}{a numeric vector}
\item{\samp{431}{X.0.983652159}}{a numeric vector}
\item{\samp{432}{X.0.99230868}}{a numeric vector}
\item{\samp{433}{X.0.981947863}}{a numeric vector}
\item{\samp{434}{X.0.972297854}}{a numeric vector}
\item{\samp{435}{X.0.99537973}}{a numeric vector}
\item{\samp{436}{X.0.997568877}}{a numeric vector}
\item{\samp{437}{X.0.984087088}}{a numeric vector}
\item{\samp{438}{X.0.994326368}}{a numeric vector}
\item{\samp{439}{X.0.985259304}}{a numeric vector}
\item{\samp{440}{X.0.999863715}}{a numeric vector}
\item{\samp{441}{X.0.999666035}}{a numeric vector}
\item{\samp{442}{X.0.999953128}}{a numeric vector}
\item{\samp{443}{X.0.990216845}}{a numeric vector}
\item{\samp{444}{X.0.994827202}}{a numeric vector}
\item{\samp{445}{X.0.99441158}}{a numeric vector}
\item{\samp{446}{X.0.712402252}}{a numeric vector}
\item{\samp{447}{X.0.644842356}}{a numeric vector}
\item{\samp{448}{X.0.838992978}}{a numeric vector}
\item{\samp{449}{X.1.10}}{a numeric vector}
\item{\samp{450}{X.1.11}}{a numeric vector}
\item{\samp{451}{X.1.12}}{a numeric vector}
\item{\samp{452}{X.0.216907636}}{a numeric vector}
\item{\samp{453}{X0.093167361}}{a numeric vector}
\item{\samp{454}{X0.458656042}}{a numeric vector}
\item{\samp{455}{X0.373538461}}{a numeric vector}
\item{\samp{456}{X0.130112381}}{a numeric vector}
\item{\samp{457}{X0.273330526}}{a numeric vector}
\item{\samp{458}{X.0.086917939}}{a numeric vector}
\item{\samp{459}{X.0.496486316}}{a numeric vector}
\item{\samp{460}{X.0.779179348}}{a numeric vector}
\item{\samp{461}{X.0.99986503}}{a numeric vector}
\item{\samp{462}{X.0.999932213}}{a numeric vector}
\item{\samp{463}{X.0.999972186}}{a numeric vector}
\item{\samp{464}{X.0.999970481}}{a numeric vector}
\item{\samp{465}{X.0.99993176}}{a numeric vector}
\item{\samp{466}{X.0.999958537}}{a numeric vector}
\item{\samp{467}{X.0.99992942}}{a numeric vector}
\item{\samp{468}{X.0.999984651}}{a numeric vector}
\item{\samp{469}{X.0.999863443}}{a numeric vector}
\item{\samp{470}{X.0.999966888}}{a numeric vector}
\item{\samp{471}{X.0.999936974}}{a numeric vector}
\item{\samp{472}{X.0.999953963}}{a numeric vector}
\item{\samp{473}{X.0.99986442}}{a numeric vector}
\item{\samp{474}{X.0.999961892}}{a numeric vector}
\item{\samp{475}{X.0.999568511}}{a numeric vector}
\item{\samp{476}{X.0.999978068}}{a numeric vector}
\item{\samp{477}{X.0.999991587}}{a numeric vector}
\item{\samp{478}{X.0.999990207}}{a numeric vector}
\item{\samp{479}{X.0.99996861}}{a numeric vector}
\item{\samp{480}{X.0.999807285}}{a numeric vector}
\item{\samp{481}{X.0.998346118}}{a numeric vector}
\item{\samp{482}{X.0.99896122}}{a numeric vector}
\item{\samp{483}{X.0.99961866}}{a numeric vector}
\item{\samp{484}{X.0.999989344}}{a numeric vector}
\item{\samp{485}{X.0.999935814}}{a numeric vector}
\item{\samp{486}{X.0.998388009}}{a numeric vector}
\item{\samp{487}{X.0.999642684}}{a numeric vector}
\item{\samp{488}{X.0.99997268}}{a numeric vector}
\item{\samp{489}{X.0.999969823}}{a numeric vector}
\item{\samp{490}{X.0.999975576}}{a numeric vector}
\item{\samp{491}{X.0.999905778}}{a numeric vector}
\item{\samp{492}{X.0.999986109}}{a numeric vector}
\item{\samp{493}{X.0.999937118}}{a numeric vector}
\item{\samp{494}{X.0.999751152}}{a numeric vector}
\item{\samp{495}{X.0.999072274}}{a numeric vector}
\item{\samp{496}{X.0.999927536}}{a numeric vector}
\item{\samp{497}{X.0.99996574}}{a numeric vector}
\item{\samp{498}{X.0.999909203}}{a numeric vector}
\item{\samp{499}{X.0.999892923}}{a numeric vector}
\item{\samp{500}{X.0.99944433}}{a numeric vector}
\item{\samp{501}{X.0.999957631}}{a numeric vector}
\item{\samp{502}{X.0.99995875}}{a numeric vector}
\item{\samp{503}{X.0.951549871}}{a numeric vector}
\item{\samp{504}{X.0.966537569}}{a numeric vector}
\item{\samp{505}{X.0.948523894}}{a numeric vector}
\item{\samp{506}{X.0.981574846}}{a numeric vector}
\item{\samp{507}{X.0.921602884}}{a numeric vector}
\item{\samp{508}{X.0.951549871.1}}{a numeric vector}
\item{\samp{509}{X.0.998989478}}{a numeric vector}
\item{\samp{510}{X.0.972006561}}{a numeric vector}
\item{\samp{511}{X.0.646376446}}{a numeric vector}
\item{\samp{512}{X.0.842105263}}{a numeric vector}
\item{\samp{513}{X.0.061629034}}{a numeric vector}
\item{\samp{514}{X.0.446120233}}{a numeric vector}
\item{\samp{515}{X.0.797046006}}{a numeric vector}
\item{\samp{516}{X.0.993610272}}{a numeric vector}
\item{\samp{517}{X.0.994226058}}{a numeric vector}
\item{\samp{518}{X.0.992838651}}{a numeric vector}
\item{\samp{519}{X.0.993778052}}{a numeric vector}
\item{\samp{520}{X.0.988171896}}{a numeric vector}
\item{\samp{521}{X.0.993610272.1}}{a numeric vector}
\item{\samp{522}{X.0.999918116}}{a numeric vector}
\item{\samp{523}{X.0.991735905}}{a numeric vector}
\item{\samp{524}{X.1.13}}{a numeric vector}
\item{\samp{525}{X.0.936507937}}{a numeric vector}
\item{\samp{526}{X0.349260344}}{a numeric vector}
\item{\samp{527}{X.0.517126845}}{a numeric vector}
\item{\samp{528}{X.0.801005832}}{a numeric vector}
\item{\samp{529}{X.0.980134852}}{a numeric vector}
\item{\samp{530}{X.0.961300846}}{a numeric vector}
\item{\samp{531}{X.0.974128989}}{a numeric vector}
\item{\samp{532}{X.0.956012879}}{a numeric vector}
\item{\samp{533}{X.0.989893752}}{a numeric vector}
\item{\samp{534}{X.0.980134852.1}}{a numeric vector}
\item{\samp{535}{X.0.999240352}}{a numeric vector}
\item{\samp{536}{X.0.992673389}}{a numeric vector}
\item{\samp{537}{X.0.701291408}}{a numeric vector}
\item{\samp{538}{X.1.14}}{a numeric vector}
\item{\samp{539}{X.0.132479997}}{a numeric vector}
\item{\samp{540}{X0.565697089}}{a numeric vector}
\item{\samp{541}{X0.363477518}}{a numeric vector}
\item{\samp{542}{X.0.991994064}}{a numeric vector}
\item{\samp{543}{X.0.990877165}}{a numeric vector}
\item{\samp{544}{X.0.990169195}}{a numeric vector}
\item{\samp{545}{X.0.992520987}}{a numeric vector}
\item{\samp{546}{X.0.991044002}}{a numeric vector}
\item{\samp{547}{X.0.991994064.1}}{a numeric vector}
\item{\samp{548}{X.0.999936756}}{a numeric vector}
\item{\samp{549}{X.0.990536925}}{a numeric vector}
\item{\samp{550}{X.0.871305805}}{a numeric vector}
\item{\samp{551}{X.1.15}}{a numeric vector}
\item{\samp{552}{X.0.012235894}}{a numeric vector}
\item{\samp{553}{X.0.314848353}}{a numeric vector}
\item{\samp{554}{X.0.713307812}}{a numeric vector}
\item{\samp{555}{X.0.112754341}}{a numeric vector}
\item{\samp{556}{X0.030400372}}{a numeric vector}
\item{\samp{557}{X.0.464761386}}{a numeric vector}
\item{\samp{558}{X.0.018445884}}{a numeric vector}
\item{\samp{559}{X.0.841558511}}{a numeric vector}
\item{\samp{560}{X0.179912811}}{a numeric vector}
\item{\samp{561}{X.0.051718416}}{a numeric vector}
}}
\source{
{
\url{https://archive.ics.uci.edu/ml/datasets/Human+Activity+Recognition+Using+Smartphones}}

1. Davide Anguita, Alessandro Ghio, Luca Oneto, Xavier Parra and Jorge L. Reyes-Ortiz. A Public Domain Dataset for Human Activity Recognition Using Smartphones. 21th European Symposium on Artificial Neural Networks, Computational Intelligence and Machine Learning, ESANN 2013. Bruges Belgium 24-26 April 2013.
}
\usage{
HAPT
}
\description{
This sample data is drawn from the Human Activity Recognition (HAPT) Dataset on the UCI Machine Learning Repository. The original dataset includes sensor recordings from 30 subjects performing a range of daily activities. The data has 561 attributes, consisting of different accelerometer and gyroscope measurements. A small portion of the overall dataset (10 seconds of recordings for subject A) is included in this package.
}
\keyword{datasets}
